requirements.txt
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
Django==2.2.7
django-extensions==2.2.5
djangorestframework==3.10.3
Faker==2.0.4
idna==2.8
importlib-metadata==0.23
more-itertools==7.2.0
packaging==19.2
pkg-resources==0.0.0
pluggy==0.13.1
py==1.8.0
pyparsing==2.4.5
pytest==5.3.1
pytest-django==3.7.0
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
text-unidecode==1.3
urllib3==1.25.7
wcwidth==0.1.7
zipp==0.6.0
